hero: fix projection matrix
We use the viewport transform so that shorted axis is -1 to +1, this
preserves aspect ratio. Then the frustrum is very simple square (-1 to
+1 everywhere) and we can drop a lot of the FOV and aspect ratio issues.
By default we use the equivalent of FOV=90 by having the near clip plane
be the same distance as left,right and top,bottom.

This patch also enables depth buffer testing.