hero: enable backface culling
Change order of sector vertices to make backface culling work.

Sectors must be specified in counter-clockwise order, so that the
routine that generates triangles for the walls will draw the inside
facing parts of the wall in clockwise order. This permits the outside
faces of the walls to be culled.